+++ src/main/java/net/avdw/skilltracker/match/MatchMapper.java
@@ -24,6 +24,7 @@ public interface MatchMapper {
     }
 
     @Mappings({
+            @Mapping(target = MatchTable.PK, ignore = true),
             @Mapping(target = MatchTable.PLAY_DATE, expression = "java(new Date())")
     })
     MatchTable toMatchTable(GameTable gameTable, PlayerTable playerTable, Rating rating);