+++ CHANGELOG.md
@@ -3,3 +3,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- Discord interface
+- Commandline interface
+
+#### Game
+- Register a game
+- View the details of a game
+- Remove a game
+- List games
+
+#### Match
+- Add a match
+- Remove a match
+- List matches
+- Determine the quality of a match
+- Calculate the quality of a match for every team combination
+
+#### Player
+- View the details of a player
+- List players